+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+use TomatoPHP\FilamentLanguageSwitcher\Http\Controllers\LanguageController;
 
 /*
 |--------------------------------------------------------------------------
@@ -13,6 +14,6 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/languages/switcher', [\TomatoPHP\FilamentLanguageSwitcher\Http\Controllers\LanguageController::class, 'index'])
+Route::get('/languages/switcher', [LanguageController::class, 'index'])
     ->name('languages.switcher')
     ->middleware(config('filament-language-switcher.language_switcher_middlewares'));