+++ src/Console/FilamentLanguageSwitcherInstall.php
@@ -3,12 +3,9 @@
 namespace TomatoPHP\FilamentLanguageSwitcher\Console;
 
 use Illuminate\Console\Command;
-use TomatoPHP\ConsoleHelpers\Traits\RunCommand;
 
 class FilamentLanguageSwitcherInstall extends Command
 {
-    use RunCommand;
-
     /**
      * The name and signature of the console command.
      *
@@ -23,21 +20,15 @@ class FilamentLanguageSwitcherInstall extends Command
      */
     protected $description = 'install package and publish assets';
 
-    public function __construct()
-    {
-        parent::__construct();
-    }
-
     /**
      * Execute the console command.
-     *
-     * @return mixed
      */
-    public function handle()
+    public function handle(): int
     {
-        $this->info('Publish Vendor Assets');
-        $this->artisanCommand(['migrate']);
-        $this->artisanCommand(['filament:optimize']);
+        $this->info('Running migrations');
+        $this->call('migrate', ['--force' => true]);
         $this->info('Filament Language Switcher installed successfully.');
+
+        return self::SUCCESS;
     }
 }