+++ src/Traits/InteractsWithLanguages.php
@@ -3,16 +3,17 @@
 namespace TomatoPHP\FilamentLanguageSwitcher\Traits;
 
 use Illuminate\Database\Eloquent\Relations\MorphOne;
+use TomatoPHP\FilamentLanguageSwitcher\Models\UserLanguage;
 
 trait InteractsWithLanguages
 {
     public function lang(): MorphOne
     {
-        return $this->morphOne('TomatoPHP\FilamentLanguageSwitcher\Models\UserLanguage', 'model');
+        return $this->morphOne(UserLanguage::class, 'model');
     }
 
     public function getLangAttribute()
     {
-        return $this->lang()->first()?->lang ?? 'en';
+        return $this->lang()->first()?->lang ?? config('app.locale', 'en');
     }
 }