+++ tests/src/InstallCommandTest.php
@@ -1,11 +1,12 @@
 <?php
 
 use Illuminate\Support\Facades\Artisan;
+use Illuminate\Support\Facades\Schema;
 
 it('check install command', function () {
     Artisan::call('filament-language-switcher:install');
 
-    $schema = \Illuminate\Support\Facades\Schema::hasTable('user_languages');
+    $schema = Schema::hasTable('user_languages');
 
     expect($schema)->toBeTrue();
 });